--- src/services/boards/readBoards.js
@@ -1,5 +1,0 @@
-const { boardsRepository } = require('../../repository/database');
-
-const readBoards = () => boardsRepository;
-
-module.exports = readBoards;
