+++ src/services/boards/readBoards.ts
@@ -1,0 +1,6 @@
+/* eslint-disable import/no-import-module-exports */
+import { boardsRepository } from '../../repository/database';
+
+const readBoards = () => boardsRepository;
+
+module.exports = readBoards;
