--- src/services/tasks/readTask.js
@@ -1,8 +1,0 @@
-const { tasksRepository } = require('../../repository/database');
-
-const readBoard = (id) => {
-  const task = tasksRepository.find((item) => item.id === id);
-  return task;
-};
-
-module.exports = readBoard;
