+++ src/services/tasks/readTask.ts
@@ -1,0 +1,9 @@
+/* eslint-disable import/no-import-module-exports */
+import { tasksRepository } from '../../repository/database';
+
+const readBoard = (id: string) => {
+  const task = tasksRepository.find((item) => item.id === id);
+  return task;
+};
+
+module.exports = readBoard;
