--- src/services/tasks/readTasks.js
@@ -1,5 +1,0 @@
-const { tasksRepository } = require('../../repository/database');
-
-const readTasks = () => tasksRepository;
-
-module.exports = readTasks;
