+++ src/services/tasks/readTasks.ts
@@ -1,0 +1,6 @@
+/* eslint-disable import/no-import-module-exports */
+import { tasksRepository } from '../../repository/database';
+
+const readTasks = () => tasksRepository;
+
+module.exports = readTasks;
